change local state to storage state
