+++ src/reducers/books.js
@@ -1,4 +1,25 @@
-const books = (state = {}, action) => {
+const newID = () => parseInt(Math.random() * 100, 10);
+const initialState = {
+  books: [
+    {
+      id: newID(),
+      title: 'The lord of the rings',
+      category: 'Sci-Fi',
+    },
+    {
+      id: newID(),
+      title: 'Nightmare on Stret Elmer',
+      category: 'Horror',
+    },
+    {
+      id: newID(),
+      title: 'Im Watching you',
+      category: 'Suspense',
+    },
+  ],
+};
+
+const books = (state = initialState, action) => {
   const newState = { ...state };
   switch (action.type) {
     case 'CREATE_BOOK':
